requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
Werkzeug==2.0.1
bcrypt==3.2.0
PyJWT==2.1.0
